% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_auth_token.R
\name{fetch_auth_token}
\alias{fetch_auth_token}
\title{Fetch Oauth2 token to access the Bloom Credit API.}
\usage{
fetch_auth_token(audience)
}
\arguments{
\item{audience}{\code{character} Must be one of: "dev" or "prod", corresponding to the Audience Parameter.
See more here: https://developers.bloomcredit.io/docs/environments-1}
}
\value{
\code{character} Oauth2 token used to access the Bloom Credit API.
}
\description{
Fetch Oauth2 token to access the Bloom Credit API.
}
\examples{
\dontrun{
auth_token <- fetch_auth_token(audience = "dev")
}

}
