% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credit_data.R
\name{get_credit_data}
\alias{get_credit_data}
\title{Get credit report data based on \code{order_id}.}
\usage{
get_credit_data(url = NULL, order_id, auth_token)
}
\arguments{
\item{url}{\code{character} URL to endpoint.}

\item{order_id}{\code{character} Order ID returned by \link{order_credit_data}.}

\item{auth_token}{\code{character} Oauth2 token returned by \link{fetch_auth_token}.}
}
\value{
\code{list} of data.frames containing credit data report information.
}
\description{
Get credit report data based on \code{order_id}.
}
\examples{
\dontrun{
credit_data <- get_credit_data(
  audience = "dev",
  order_id = "your_order_id",
  auth_token = "your_auth_token"
)
}
}
