% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decision.R
\name{get_decision}
\alias{get_decision}
\title{Run decision model.}
\usage{
get_decision(credit_data)
}
\arguments{
\item{credit_data}{\code{list} retned from \link{get_credit_data}.}
}
\value{
\code{character} decision from model based on \code{credit_data}.
}
\description{
Run decision model.
}
\examples{
\dontrun{
get_decision(credit_data = credit_data)
}

}
