% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_utils.R
\name{make_registration_body}
\alias{make_registration_body}
\title{Make body for API call to register a consumer.}
\usage{
make_registration_body(consumer_info)
}
\arguments{
\item{consumer_info}{\code{list} containing the required named fields:
\itemize{
\item \code{first_name}:
\item \code{last_name}:
\item \code{city}:
\item \code{line1}:
\item \code{state_code}:
\item \code{zipcode}: (character) must be exactly 5 digits, e.g.: "90210".
\item \code{address_primary}: (boolean) must be \code{TRUE} or \code{FALSE}.
}}
}
\value{
\code{list} body of registration request.
}
\description{
Make body for API call to register a consumer.
}
\keyword{internal}
