% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_credit_data.R
\name{order_credit_data}
\alias{order_credit_data}
\title{Order Credit Data}
\usage{
order_credit_data(
  url = set_order_credit_data_url("dev"),
  consumer_id,
  portfolio_id,
  sku,
  auth_token
)
}
\arguments{
\item{url}{\code{character} URL to endpoint.}

\item{consumer_id}{\code{character} Consumer ID.}

\item{portfolio_id}{\code{character} Portfolio ID.}

\item{sku}{\code{character} SKU enabled for your account. E.g: "equifax-gold-soft-fico-internet".}

\item{auth_token}{\code{character} Oauth2 token returned by \link{fetch_auth_token}.}
}
\value{
Response from placing a credit order.
}
\description{
Order Credit Data
}
\examples{
\dontrun{
credit_data <- order_credit_data(
  audience = "dev",
  consumer_id = "your_consumer_id",
  portfolio_id = "your_portfolio_id",
  sku = "sku_you_want_to_order",
  auth_token = "your_auth_token"
)
}

}
