% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_consumer.R
\name{register_consumer}
\alias{register_consumer}
\title{Register a consumer.}
\usage{
register_consumer(url = set_registration_url("dev"), consumer_info, auth_token)
}
\arguments{
\item{url}{\code{character} URL to endpoint.}

\item{consumer_info}{\code{list} containing the required named fields:
\itemize{
\item \code{first_name}:
\item \code{last_name}:
\item \code{city}:
\item \code{line1}:
\item \code{state_code}:
\item \code{zipcode}: (character) must be exactly 5 digits, e.g.: "90210".
\item \code{address_primary}: (boolean) must be \code{TRUE} or \code{FALSE}.
}}

\item{auth_token}{\code{character} Oauth2 token returned by \link{fetch_auth_token}.}
}
\value{
\verb{console message} Confirmation or error from API response.
}
\description{
Register a consumer.
}
\examples{

\dontrun{
consumer_info <- list(
  ssn = "123456789",
  city = "Scranton",
  line1 = "1725 Slough Avenue",
  state_code = "PA",
  zipcode = "18503",
  date_of_birth = "1964-03-15",
  first_name = "Michael",
  last_name = "Scott",
  address_primary = TRUE
  )

auth_token <- fetch_auth_token(
  audience = "dev",
  client_id = "your_client_id",
  client_secret = "your_client_secret"
  )

 register_consumer(
  audience = "dev",
  consumer_info = consumer_info,
  auth_token = auth_token
 )

}

}
