% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_utils.R
\name{set_client_id}
\alias{set_client_id}
\title{Set your \code{client_id} using \code{options()}.}
\usage{
set_client_id(client_id)
}
\arguments{
\item{client_id}{\code{character} Your Bloom Credit API \code{client_id}.}
}
\description{
Set your \code{client_id} using \code{options()}.
}
