% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_utils.R
\name{set_client_secret}
\alias{set_client_secret}
\title{Set your \code{client_secret} using \code{options()}.}
\usage{
set_client_secret(client_secret)
}
\arguments{
\item{client_secret}{\code{character} Your Bloom Credit API \code{client_secret}.}
}
\description{
Set your \code{client_secret} using \code{options()}.
}
