% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_utils.R
\name{set_order_credit_data_url}
\alias{set_order_credit_data_url}
\title{Set the API URL based on audience to order credit data.}
\usage{
set_order_credit_data_url(audience)
}
\arguments{
\item{audience}{\code{character} Must be one of: "dev" or "prod", corresponding to the Audience Parameter.
See more here: https://developers.bloomcredit.io/docs/environments-1}
}
\value{
\code{character} URL for consumer registration endpoint.
}
\description{
Set the API URL based on audience to order credit data.
}
\examples{
\dontrun{
order_credit_data_url <- set_order_credit_data_url("dev")
}

}
