% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_token_url}
\alias{set_token_url}
\title{Set the API URL based on audience.}
\usage{
set_token_url(audience)
}
\arguments{
\item{audience}{\code{character} Must be one of: "dev" or "prod", corresponding to the Audience Parameter.
See more here: https://developers.bloomcredit.io/docs/environments-1}
}
\value{
\code{character} URL for Oauth2 token API endpoint.
}
\description{
Set the API URL based on audience.
}
\examples{
url <- set_url("dev")

}
\keyword{internal}
