% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloomR-package.R
\docType{package}
\name{bloomR-package}
\alias{bloomR}
\alias{bloomR-package}
\title{bloomR: Access the Bloom Credit API using R}
\description{
Access the Bloom Credit API using R to place orders, retrieve data, and run a basic decisioning model.
}
\author{
\strong{Maintainer}: Matt Roumaya \email{matthewroumaya@gmail.com} (\href{https://orcid.org/0000-0002-8009-2771}{ORCID})

}
\keyword{internal}
