% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_auth_token.R
\name{fetch_auth_token}
\alias{fetch_auth_token}
\title{Fetch Oauth2 token to access the Bloom Credit API.}
\usage{
fetch_auth_token(
  url = set_token_url("dev"),
  client_id = getOption("bloom_api_client_id"),
  client_secret = getOption("bloom_api_client_secret")
)
}
\arguments{
\item{url}{\code{character} URL to endpoint.}

\item{client_id}{\code{character}}

\item{client_secret}{\code{character}}
}
\value{
\code{character} Oauth2 token used to access the Bloom Credit API.
}
\description{
Fetch Oauth2 token to access the Bloom Credit API.
}
\examples{
\dontrun{
auth_token <- fetch_auth_token(audience = "dev")
}

}
