% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_credit_data.R
\name{order_credit_data}
\alias{order_credit_data}
\title{Title}
\usage{
order_credit_data(audience, consumer_id, portfolio_id, sku, auth_token)
}
\arguments{
\item{audience}{\code{character} Must be one of: "dev" or "prod", corresponding to the Audience Parameter.
See more here: https://developers.bloomcredit.io/docs/environments-1}

\item{consumer_id}{\code{character} Consumer ID.}

\item{portfolio_id}{\code{character} Portfolio ID.}

\item{sku}{\code{character} SKU enabled for your account. E.g: "equifax-gold-soft-fico-internet".}

\item{auth_token}{\code{character} Oauth2 token returned by \link{fetch_auth_token}.}
}
\value{
Response from placing a credit order.
}
\description{
Title
}
\examples{
\dontrun{
credit_data <- order_credit_data(
  audience = "dev",
  consumer_id = "your_consumer_id",
  portfolio_id = "your_portfolio_id",
  sku = "sku_you_want_to_order",
  auth_token = "your_auth_token"
)
}

}
